Split ChartSettings files into their own components (#49281)
