+++ frontend/src/metabase/visualizations/components/ChartSettings/BaseChartSettings/index.ts
@@ -1,0 +1,1 @@
+export * from "./BaseChartSettings";
