+++ frontend/src/metabase/visualizations/components/ChartSettings/ChartSettingsFooter/ChartSettingsFooter.tsx
@@ -2,7 +2,7 @@ import { t } from "ttag";
 
 import Button from "metabase/core/components/Button";
 
-import { ChartSettingsFooterRoot } from "./ChartSettings.styled";
+import { ChartSettingsFooterRoot } from "./ChartSettingsFooter.styled";
 
 export type ChartSettingsFooterProps = {
   onDone: () => void;
