+++ frontend/src/metabase/visualizations/components/ChartSettings/DashboardChartSettings/DashboardChartSettings.unit.spec.tsx
@@ -9,7 +9,7 @@ import {
   createMockVisualizationSettings,
 } from "metabase-types/api/mocks";
 
-import { DashboardChartSettings } from "./ChartSettings";
+import { DashboardChartSettings } from "./DashboardChartSettings";
 import type { DashboardChartSettingsProps } from "./types";
 
 registerVisualizations();
