+++ frontend/src/metabase/visualizations/components/ChartSettings/QuestionChartSettings/types.ts
@@ -1,0 +1,9 @@
+import type { BaseChartSettingsProps } from "../BaseChartSettings/types";
+import type {
+  BaseChartSettingsTestProps,
+  CommonChartSettingsProps,
+} from "../types";
+
+export type QuestionChartSettingsProps = CommonChartSettingsProps &
+  Pick<BaseChartSettingsProps, "initial" | "computedSettings" | "question"> &
+  BaseChartSettingsTestProps;
